NapCat/bot插件/JM下载BOT/requirements.txt
nonebot2
nonebot-adapter-onebot
jmcomic
pillow
